📝 Modify(#196): BadgeContainerSkeleton 컴포넌트 추가 및 ProfileDetails에서 Suspense로 로딩 상태 관리
