+++ src/features/profile/ui/BadgeContainer.tsx
@@ -1,34 +1,52 @@
 import * as S from './styles';
+import { useState } from 'react';
 import { getImageUrl } from '@/utils/getImageUrl';
-import type { ChallengeItem } from '@features/user/types';
-import { Dispatch, SetStateAction } from 'react';
+import { usersChallengesQuery } from '@/features/user/queries';
+import useUserStore from '@store/useUserStore';
 import { useNavigate } from 'react-router-dom';
+import { useMediaQuery } from '@modern-kit/react';
+import { MEDIA_QUERY_MAP } from '@style/constants';
+import { isLoggedIn } from '@features/user/service/authUtils';
 
-interface BadgeContainerProps {
-  completedChallenges: ChallengeItem[];
-  page: number;
-  setPage: Dispatch<SetStateAction<number>>;
-  totalPage: number;
-}
-
-export default function BadgeContainer({
-  completedChallenges,
-  page,
-  setPage,
-  totalPage,
-}: BadgeContainerProps) {
+export default function BadgeContainer() {
   const navigate = useNavigate();
+  const [page, setPage] = useState(1);
+
+  const { user } = useUserStore();
+  const loggedIn = isLoggedIn(user);
+
+  const isMobile = useMediaQuery(MEDIA_QUERY_MAP.mobile);
+  const limit = isMobile ? 1 : 4;
+
+  if (!loggedIn) {
+    return (
+      <S.BadgeWrapper>
+        <S.EmptyBadgeContainer>
+          <p>로그인이 필요합니다.</p>
+          <S.BadgeGuideText>
+            뱃지를 보려면 먼저 로그인해주세요.
+          </S.BadgeGuideText>
+          <button onClick={() => navigate('/login')}>로그인하기</button>
+        </S.EmptyBadgeContainer>
+      </S.BadgeWrapper>
+    );
+  }
+
+  const { data: challengesData } = usersChallengesQuery.useGetChallenges({
+    page,
+    limit,
+    completed: true,
+  });
+
+  const completedChallenges = challengesData?.contents || [];
+  const totalPage = challengesData?.totalPage ?? 1;
 
   const handlePrevPage = () => {
-    if (page > 1) {
-      setPage(page - 1);
-    }
+    if (page > 1) setPage(page - 1);
   };
 
   const handleNextPage = () => {
-    if (page < totalPage) {
-      setPage(page + 1);
-    }
+    if (page < totalPage) setPage(page + 1);
   };
 
   return (
@@ -61,9 +79,7 @@ export default function BadgeContainer({
           <S.BadgeGuideText>
             도전과제를 완료하고 다양한 뱃지를 모아보세요!
           </S.BadgeGuideText>
-          <S.GoToQuestButton onClick={() => navigate('/quest')}>
-            도전과제 보러가기
-          </S.GoToQuestButton>
+          <button onClick={() => navigate('/quest')}>도전과제 보러가기</button>
         </S.EmptyBadgeContainer>
       )}
